Move RadioButtonGroup to RadioButton/Group for consistency
