+++ lib/Form/RadioButton/Group.js
@@ -4,7 +4,7 @@ import assign from 'object-assign';
 import RadioButton from '../RadioButton';
 import RadioButtonOther from '../RadioButton/Other';
 
-class RadioButtonGroup extends Component {
+export default class Group extends Component {
   static propTypes = {
     choices: PropTypes.arrayOf(PropTypes.shape({
       label: PropTypes.string.isRequired,
@@ -73,6 +73,3 @@ class RadioButtonGroup extends Component {
     );
   }
 }
-
-
-export default RadioButtonGroup;
