+++ lib/index.js
@@ -17,7 +17,7 @@ export PlanBoxWrapper from './PlanBox/Wrapper';
 export PlanBoxPricing from './PlanBox/Pricing';
 export PlanBoxAllowanceDetails from './PlanBox/AllowanceDetails';
 export Carousel from './Carousel';
-export RadioButtonGroup from './Form/RadioButtonGroup';
+export RadioButtonGroup from './Form/RadioButton/Group';
 export DropdownSwitcher from './DropdownSwitcher';
 export Modal from './Modal';
 export Field from './Field';
